requirements.txt
tensorflow
matplotlib
tqdm
Pillow
opencv-python
pyyaml
IPython